% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbl-commute-traffic-data.R
\docType{data}
\name{get_closest_stops}
\alias{get_closest_stops}
\title{Data for every Buss, Train, Tram and Ferry Service}
\format{An object of class \code{"data.frame"}.}
\source{
\href{http://example.com}{Example example}

\href{http://example.com}{Example example}

\href{http://example.com}{Example example}

\href{http://example.com}{Example example}
}
\usage{
data(services)
}
\description{
Contains data on each service (i.e. individual departures). For each service the following variables
are provided: serviceId containing an identifier id, line indicating which line the service operates,
footnoteId used to map the correct schedule to the service and direction indicating the direction
of the service.
}
\examples{
data(services)
station_index <- services$stationIndex
Location and Name Data for every Buss, Train, Tram and Ferry Stations

Contains geographical locations for the different stops. The variables provided are: stationId, an identifier id
for the stops. The vaiable stationLon prodiving the longitudinal data (in GWS84) for the stop. The vaiable
stationLat prodiving the latitudinal data (in GWS84) for the stop. The variable name providing the name for the
stop.

}
\keyword{datasets}

